requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
openai==1.35.0
langchain==0.2.6
langchain-openai==0.1.14
langchain-core==0.2.11
transformers==4.41.2
torch==2.3.1
pytest==8.2.2
httpx==0.27.0
